requirements.txt
python-frontmatter==1.0.0
Markdown==3.4.4
Jinja2==3.1.2
dotmap==1.3.30
